Reuse workdir if it already exists
Since the executables are now always run in a Docker container, we can assume
that in production the workdir is used only to store the Git repo. So if it
already exists, we no longer delete it and re-clone mathlib. This speeds up
testing since we don't have to clone mathlib for every run.