supervisor: rearchitect commit page
The per-file timing page is now a commit page which shows various data about a
single commit. This data has been removed from the main page.

The database schema has also changed: per-file timings are no longer associated
with a commit but with a build. This allows us to implement multiple timings per
commit if there's ever a need for that.